Fix Convex calendar runtime split
